feat: Fix business logic

+++ src/main/java/org/example/WinScenario.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 public @Data class WinScenario {
     private String label;
-    private int multiplier;
+    private double multiplier;
     private int counter;
     private String[] coveredArea;
 }
